pyproject.toml
[project]
name = "backendcourse"
version = "0.1.0"
description = "backend develop course"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.14.1",
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "arrow==1.3.0",
    "asyncpg==0.30.0",
    "attrs==25.1.0",
    "backoff==2.2.1",
    "bcrypt==4.0.1",
    "celery>=5.4.0",
    "certifi==2024.8.30",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.1",
    "click==8.1.7",
    "colorama==0.4.6",
    "distlib==0.3.9",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "faker==36.1.1",
    "fastapi==0.115.6",
    "fastapi-cache2>=0.2.2",
    "fastapi-cli==0.0.6",
    "filelock==3.17.0",
    "fqdn==1.5.1",
    "graphql-core==3.2.6",
    "greenlet==3.1.1",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "harfile==0.3.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "hypothesis==6.125.3",
    "hypothesis-graphql==0.11.1",
    "hypothesis-jsonschema==0.23.1",
    "identify==2.6.6",
    "idna==3.10",
    "iniconfig==2.0.0",
    "isoduration==20.11.0",
    "isort==5.13.2",
    "jinja2==3.1.4",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "junit-xml==1.9",
    "mako==1.3.8",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "multidict==6.1.0",
    "nodeenv==1.9.1",
    "orjson>=3.10.18",
    "packaging==24.2",
    "passlib==1.7.4",
    "pillow>=11.2.1",
    "platformdirs==4.3.6",
    "pluggy==1.5.0",
    "pre-commit==4.1.0",
    "pre-commit-hooks==5.0.0",
    "propcache==0.2.1",
    "pydantic==2.10.3",
    "pydantic-core==2.27.1",
    "pydantic-settings==2.7.1",
    "pygments==2.18.0",
    "pyjwt==2.10.1",
    "pyrate-limiter==3.7.0",
    "pyright>=1.1.402",
    "pytest==8.3.4",
    "pytest-asyncio>=0.26.0",
    "pytest-dotenv>=0.5.2",
    "pytest-subtests==0.14.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.19",
    "pyyaml==6.0.2",
    "redis>=6.0.0",
    "referencing==0.36.2",
    "requests==2.32.3",
    "rfc3339-validator==0.1.4",
    "rfc3987==1.3.8",
    "rich==13.9.4",
    "rich-toolkit==0.12.0",
    "rpds-py==0.22.3",
    "ruamel-yaml==0.18.10",
    "ruff==0.9.3",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "sqlalchemy==2.0.37",
    "starlette==0.41.3",
    "starlette-testclient==0.4.1",
    "tomli==2.2.1",
    "tomli-w==1.2.0",
    "typer==0.15.1",
    "types-python-dateutil==2.9.0.20241206",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "uri-template==1.3.0",
    "urllib3==2.3.0",
    "uvicorn==0.32.1",
    "uvloop==0.21.0",
    "virtualenv==20.29.1",
    "watchfiles==1.0.0",
    "webcolors==24.11.1",
    "websockets==14.1",
    "werkzeug==3.1.3",
    "yarl==1.18.3",
]
